Move Application to AppInfo
